server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
openpyxl==3.1.5
openai==1.97.1
python-dotenv==1.1.1
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9